Start plumbing MQTT notifications; restrict new member activation to valid AMP members
